a lil dotnet cleanup
